runtime: add modules
The udev module is grabbed from
https://github.com/alexlarsson/spotify-app/blob/master/com.spotify.Client.json

udev isn't stable so bringing it into a runtime is questionable.
But chromium needs it, we need to support the electron prebuilt
binaries with chromium, so I don't see an alternative.

We are building the latest stable node only for the Sdk. Since electron
uses its own copy of v8 to run its main process, we shouldn't need
node in the runtime.