Build a "base app" instead of runtime
We will build a "base app" instead of a runtime for electron app
dependencies. Essentially moving everything specified here from
/usr to /app and distributing them with the application. On its
own this base app will be pretty useless, but it can be specified
as a base for another flatpak build, a feature that just landed
in flatpak-builder.

This should be easier to maintain than an actual runtime. i.e.
freedesktop can push our a critical security fix for the platform
without us needing to rebuild this base app.

Relevant discussion here
https://github.com/flatpak/flatpak/issues/347